+++ web/src/App.tsx
@@ -22,6 +22,7 @@ import LambdaPage from './pages/LambdaPage'
 import BackupPage from './pages/BackupPage'
 import SearchPage from './pages/SearchPage'
 import MigrationPage from './pages/MigrationPage'
+import CostPage from './pages/CostPage'
 import OIDCCallbackPage from './pages/OIDCCallbackPage'
 
 export default function App() {
@@ -51,6 +52,7 @@ export default function App() {
                 <Route path="/backup" element={<BackupPage />} />
                 <Route path="/search" element={<SearchPage />} />
                 <Route path="/migrate" element={<MigrationPage />} />
+                <Route path="/cost" element={<CostPage />} />
               </Route>
             </Route>
           </Routes>